重複処理の統一化
